list_models.py
import os
import google.generativeai as genai
from dotenv import load_dotenv
load_dotenv()
api_key = os.getenv('GOOGLE_API_KEY')
if not api_key:
    raise ValueError('GOOGLE_API_KEY not set')

genai.configure(api_key=api_key)

models = genai.list_models()
print('Available models:')
for m in models:
    print(m.name)
