requirements.txt
streamlit
requests
pydantic
python-dotenv
langchain
langchain-community
sentence-transformers
pinecone
tiktoken
firecrawl-py
google-genai
google-generativeai
